feat: Refactor getNodeText function into own file
